需求[#18704]: 运行监控-系统监控-获取cpu信息接口开发
    [描述说明]:运行监控-系统监控-获取cpu信息接口开发
    [问题原因]:
    [解决方法]:
    [测试建议]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/device/service/SystemInfoService.java
@@ -29,7 +29,7 @@ public class SystemInfoService {
     public List<Map<String,String>> getCPU() throws Exception {
         List<Map<String,String>> result = systemInfoUtil.getCpuUsage();
         for (Map<String,String> item: result) {
-            if (0L == Long.parseLong(item.get("cpu"))){
+            if (0 == Double.parseDouble(item.get("cpu"))){
                     result.remove(item);
             }
         }
@@ -40,7 +40,7 @@ public class SystemInfoService {
     public List<Map<String,String>> getSwap() throws Exception {
         List<Map<String,String>> result = systemInfoUtil.getDeskUsage();
         for (Map<String,String> item: result) {
-            if (0L == Long.parseLong(item.get("all"))){
+            if (0 == Double.parseDouble(item.get("all"))){
                 result.remove(item);
             }
         }
